+++ src/sys/device/display/x11/error.rs
@@ -1,4 +1,4 @@
-// devela::sys::display::x11::error
+// devela::sys::device::display::x11::error
 //
 //! Defines [`XError`].
 //
@@ -7,7 +7,7 @@ use crate::{Display, Error, FmtResult, Formatter};
 
 #[doc = crate::_tags!(linux error)]
 /// XCB/X11 error categories.
-#[doc = crate::_doc_location!("sys/display/x11")]
+#[doc = crate::_doc_location!("sys/device/display/x11")]
 ///
 /// Represents the main classes of failures that can occur when interacting with
 /// the X server through XCB. These variants cover connection issues, setup failures,