+++ src/sys/device/display/x11/xkb.rs
@@ -1,4 +1,4 @@
-// devela::sys::display::x11::xkb
+// devela::sys::device::display::x11::xkb
 //
 //! Defines [`KeyRepeatFilter`], [`XkbInfo`], [`XkbState`].
 //
@@ -11,6 +11,7 @@ use super::{XError, raw};
 use crate::{Bitwise, Key, KeyDead, KeyMod, KeyMods, KeyPad, KeyState, is};
 
 /// Tracks the minimal state needed to classify `Press` vs `Repeat`.
+#[doc = crate::_doc_location!("sys/device/display/x11")]
 ///
 /// This is stored in [`XDisplay`][crate::XDisplay], and used in [`crate::XEvent::to_event_key`].
 #[derive(Clone, Copy, Debug, Default, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -52,6 +53,7 @@ impl KeyRepeatFilter {
 }
 
 /// Intermediate result combining semantic + physical keys + modifiers.
+#[doc = crate::_doc_location!("sys/device/display/x11")]
 ///
 /// This is produced by [`XkbState::translate_key`] and consumed by [`crate::XEvent::to_event_key]`.
 #[derive(Debug)]
@@ -122,6 +124,7 @@ impl XkbKeyInfo {
 // }
 
 /// XKB translation state.
+#[doc = crate::_doc_location!("sys/device/display/x11")]
 ///
 /// Provides keysym interpretation, physical-key identification,
 /// and modifier mapping for X11 keyboard events.