+++ src/sys/device/midi/mod.rs
@@ -1,0 +1,9 @@
+// devela::sys::device::midi
+//
+//! Midi functionality.
+//
+
+crate::structural_mods! { // _mods
+    _mods {
+    }
+}