+++ src/sys/hw/mod.rs
@@ -3,12 +3,26 @@
 #![doc = crate::_DOC_SYS_HW!()]
 //
 
-mod audio; //
+// mod camera; //
+// mod hid; // Evdev* WIP
+// mod opengl; //
+// mod scan; //
+// mod sensor; //
+// mod serial; //
+// mod storage; //
+// mod usb; //
 
 crate::structural_mods! { // _mods
     _mods {
-        pub use super::{
-            audio::_all::*,
-        };
+        // pub use super::hid::_all::*; // WIP
+        // pub use super::{
+        //     camera::_all::*,
+        //     opengl::_all::*,
+        //     scan::_all::*,
+        //     sensor::_all::*,
+        //     serial::_all::*,
+        //     storage::_all::*,
+        //     usb::_all::*,
+        // };
     }
 }